pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.116.1",
    "flask>=3.1.1",
    "gunicorn>=23.0.0",
    "jinja2>=3.1.6",
    "openai>=1.95.1",
    "pandas>=2.3.1",
    "plotly>=6.2.0",
    "psycopg2-binary>=2.9.10",
    "python-multipart>=0.0.20",
    "python-telegram-bot[webhooks]==21.5",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "sqlalchemy>=2.0.41",
    "streamlit>=1.46.1",
    "uvicorn>=0.35.0",
    "werkzeug>=3.1.3",
]
